*** empty log message ***
git-svn-id: svn://svn.code.sf.net/p/iup/iup/trunk@155 2f08bc20-984a-4df8-b195-9b5d5d477537

+++ iup/src/iup_maskparse.c
@@ -438,7 +438,7 @@ static int iMaskParseFactor (ImaskParseVars * vars)
         break;
 
       default:
-        if (isdigit (vars->string[vars->j]))
+        if (isdigit((int)vars->string[vars->j]))
         {
           sscanf (&vars->string[vars->j], "%3d", &temp);
           if (temp > 255)
